% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcreml.R
\name{fitCReml}
\alias{fitCReml}
\title{Fit a spatial model by Composite REML}
\usage{
fitCReml(form, data, dS, coordsMat, init, pointsj, pointsjpair, pointsk,
  cov.model = cov.model, lower = NULL)
}
\arguments{
\item{form}{a formula that specifies the response and the fixed effect variables.}

\item{data}{a data.frame that contains the response variable and the fixed effect variables.}

\item{dS}{a list of functions to compute the first derivative of Sigma (dSigma/dtheta).}

\item{coordsMat}{a N by D coordinate matrix. D being the spatial dimension, typically 2 or 3.}

\item{init}{named vector of initial value for optimiation of \code{recl}. Must (at least) contain named elements \emph{sigma2} and \emph{phi}. Can also contain \emph{nugget} and \emph{kappa}.}

\item{pointsjpair}{A list of pair integers from \code{1} to \code{length(pointsj)}. Each pair indicates which pairs of \code{pointjs} components should be bundled together in the variance calculation. See Details.}

\item{cov.model}{character string that specifies the spatial covariance model. See \code{?geoR::cov.spatial}}

\item{lower}{Numeric vector of lower bound for the estimating parameters. Default to NULL (no lower bound).}

\item{poinstj}{A list of row numbers that indicates which row belongs to block j (block j is itself a union of block k and block l). See Details.}
}
\description{
Fit a spatial model by Composite REML
}

