% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSigma.R
\name{genSigma}
\alias{genSigma}
\title{Generate the Covariance matrix Sigma.}
\usage{
genSigma(distM, theta, cov.model, addNugget = TRUE)
}
\arguments{
\item{distM}{Either a distance matrix (result from \code{dist}), or a matrix of distances.}

\item{theta}{A vector of named covariance paramter. Must (at least) contain named element  \emph{sigma2} and \emph{phi}. Other possible elements include \emph{nugget} and \emph{kappa}.}

\item{cov.model}{Argument that is passed to \code{geoR::cov.spatial}.}

\item{addNugget}{TRUE/FALSE. If true nugget is added to the covariance matrix.}
}
\description{
\code{genSigma} generates the covariance matrix.
}

