% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad.R
\name{grHs}
\alias{grHs}
\title{Computes the grad and hessian of \code{recl}.}
\usage{
grHs(theta, dS, coordsMat, Y, Xmat, pointsj, cov.model = "exp", gr = TRUE,
  hs = TRUE, ...)
}
\arguments{
\item{theta}{named vector of covariance parameter. Must (at least) contain named elements \emph{sigma2} and \emph{phi}. Can also contain \emph{nugget} and \emph{kappa}.}

\item{dS}{a list of functions to compute the first derivative of Sigma (dSigma/dtheta).}

\item{coordsMat}{a N by D coordinate matrix. D being the spatial dimension, typically 2 or 3.}

\item{Y}{the response vector}

\item{Xmat}{the design matrix (not the data.frame)}

\item{cov.model}{character string that specifies the spatial covariance model. See \code{?geoR::cov.spatial}}

\item{gr}{TRUE/FALSE. Whether the grad vector should be returned.}

\item{hs}{TRUE/FALSE. Whether the Hessian matrix should be returned.}

\item{poinstj}{A list of row numbers that indicates which row belongs to block j (block j is itself a union of block k and block l). See Details.}
}
\description{
Computes the grad and hessian of \code{recl}.
}

