% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recl.R
\name{recl}
\alias{recl}
\title{Calculate the Residual Composite log-Likelihood.}
\usage{
recl(theta, coordsMat, Y, Xmat, pointsj, cov.model = "exp", ...)
}
\arguments{
\item{theta}{A named numeric vector containing the covariance parameters. It should (at least) have element \emph{sigma2} and \emph{phi}. Other possible elements are \emph{nugget} and \emph{kappa}.}

\item{coordsMat}{A N by D matrix of coordinates. D is the spatial dimension, typically 2 or 3.}

\item{Y}{the response vector}

\item{Xmat}{the design matrix (not the data.frame)}

\item{pointsj}{A list. Each list contains the row number of the data matrix (Y or Xmat or coordsMat) that correspond to the (k,l) block.}

\item{cov.model}{This is passed to \code{geoR::cov.spatial}}
}
\description{
\code{recl} calculates the residual composite log-likelihood.
}

